WIP: Add CGB colour palette support
This is used by CGB when DMG-only games are played on it.

Additionally changed the framebuffer colour palette to RGB555 from
RGB565, since this is the same number of bits the CGB used.

Note that peanut-gb is still DMG only. I'm just squeezing as much colour
as I can without going full CGB.

Signed-off-by: Mahyar Koshkouei <mahyar.koshkouei@gmail.com>