GB: use mask instead of modulo for banking wrap
This is an optimisation for low powered microcontrollers that do not
have a modulo CPU instruction. A bit-mask is used to wrap the ROM bank
instead of performing a modulo.

This additionally removes support for 1.1MiB, 1.2MiB, and 1.5MiB ROMs.
There are no known ROM files that use these sizes.

Signed-off-by: Mahyar Koshkouei <mk@deltabeard.com>