peanut: white frame after LCD enable
This is a partial fix for issue #33, whereby the Game Boy LCD must
display a white screen for one frame after the LCD is enabled. This
stops garbled graphics being displayed by some games when the LCD is
enabled.

This is only a "partial" fix because Peanut-GB will simply not update
the LCD when the LCD has been enabled by the game. This means that
instead of showing a white screen, it is possible that the last
previously rendered frame will remain on the screen. This is not
accurate.

Signed-off-by: Mahyar Koshkouei <mk@deltabeard.com>