gb: improve gb_error documentation
It is now clear the gb_error() must never return. Returning from
gb_error is still undefined as it was before, but now the compiler is
told to not expect gb_error to return.

The value passed to gb_error() is now always the address with which the
error occured. Each call of gb_error() has a comment explaining what
address is passed.

Closes issue #19.

Signed-off-by: Mahyar Koshkouei <mk@deltabeard.com>