gb: fix bootrom being re-enabled
Fixed issue #110 whereby a spurious write could enable the bootrom.
Previously, the bootrom could be enabled by writing 0 to 0xFF50. But
this is not possible on real hardware after the bootrom writes 1 to
0xFF50. No game is expected to attempt to re-enable the bootrom, but a
bug in a game or the emulator could cause it to happen.

If the bootrom was enabled by the game, and the gb_bootrom_read function
was set to NULL, then a NULL pointer exception was possible.

Signed-off-by: Mahyar Koshkouei <mk@deltabeard.com>