gb: fix infinite loop on infinite halt
Fixes issue #83 whereby the Game Boy may be placed into HALT mode
indefinitely, which is a valid state to be in (e.g. when Dragon Quest
III puts a message on the screen forever). Peanut-GB now returns on each
VBLANK (which means after rendering each frame), to allow the emulation
to continue.

Signed-off-by: Mahyar Koshkouei <mk@deltabeard.com>