Optimise the sprite sorting by not realy sorting.
I'm not sure if this consumes less memory or if its even faster but i can say for sure that it will use way less stack space. It might be intresting change to defualt entry positioning but anything besides front and, the current, back will reguire a diffrent loop and its also pretty game specific.

Signed-off-by: Quentin Boswank <qubos@outlook.de>