lcd: count time LCD is off
This fixes an issue whereby Peanut-GB would continuously step the CPU
whilst the LCD was switched off. gb->gb_frame is now set when a certain
period of time has passed rather than when a new frame is drawn.

Tested working in RedstarBluestar (the start screen is now delayed by a
short time).

Closes #89.

Signed-off-by: Mahyar Koshkouei <mk@deltabeard.com>